git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@787 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/main/scala/cspfj/constraint/Constraint.scala
@@ -156,7 +156,7 @@ trait Constraint extends Weighted with Identified with IOBinomialHeapNode[Constr
     c
   }
 
-  def getEvaluation: Double
+  def getEvaluation: Int
 
   //def tupleManager: TupleManager
