+++ src/main/scala/cspfj/priorityqueues/Key.scala
@@ -1,7 +1,7 @@
 package cspfj.priorityqueues
 
 trait Key[T] extends Ordering[T] {
-  def getKey(o: T): Double
+  def getKey(o: T): Int
 
   def compare(x: T, y: T) = getKey(x).compareTo(getKey(y))
 }
